requirements.txt
# requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
sqlalchemy==2.0.17
python-dotenv==1.0.0
pydantic==2.0.2
pydantic-settings==2.0.1
alembic==1.11.1
jinja2==3.1.2